+++ include/mirai/events/MemberCardChangeEvent.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#ifndef mirai_cpp_events_group_name_change_hpp_H_
-#define mirai_cpp_events_group_name_change_hpp_H_
+#ifndef mirai_cpp_events_MemberCardChangeEvent_hpp_H_
+#define mirai_cpp_events_MemberCardChangeEvent_hpp_H_
 
 #include "mirai/third-party/nlohmann/json.hpp"
 #include "event_interface.hpp"
@@ -10,19 +10,19 @@
 namespace Cyan
 {
 	/**
-	 * \brief 群名称被改变
+	 * \brief 群名片修改事件(由于服务器并不会告知名片变动, 此事件只能由 mirai 在发现变动时才广播. 不要依赖于这个事件.)
 	 */
-	class GroupNameChangeEvent : public EventBase
+	class MemberCardChangeEvent : public EventBase
 	{
 	public:
 		string OriginName;
 		string CurrentName;
-		Group_t Group;
+		GroupMember Member;
 		GroupMember Operator;
 
 		static MiraiEvent GetMiraiEvent()
 		{
-			return MiraiEvent::GroupNameChangeEvent;
+			return MiraiEvent::MemberCardChangeEvent;
 		}
 
 		bool OperatorIsBot() const
@@ -34,7 +34,7 @@ namespace Cyan
 		{
 			this->OriginName = j["origin"].get<string>();
 			this->CurrentName = j["current"].get<string>();
-			this->Group.Set(j["group"]);
+			this->Member.Set(j["member"]);
 			if (!j["operator"].is_null())
 			{
 				this->Operator.Set(j["operator"]);
@@ -45,10 +45,10 @@ namespace Cyan
 		virtual json ToJson() const override
 		{
 			json j = json::object();
-			j["type"] = "GroupNameChangeEvent";
+			j["type"] = "MemberCardChangeEvent";
 			j["origin"] = this->OriginName;
 			j["current"] = this->CurrentName;
-			j["group"] = this->Group.ToJson();
+			j["member"] = this->Member.ToJson();
 			if (!operator_is_null_)
 				j["operator"] = this->Operator.ToJson();
 			else
@@ -62,4 +62,4 @@ namespace Cyan
 
 }
 
-#endif // !mirai_cpp_events_group_name_change_hpp_H_
+#endif
