+++ include/mirai/events/MemberJoinEvent.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#ifndef mirai_cpp_events_member_join_event_hpp_H_
-#define mirai_cpp_events_member_join_event_hpp_H_
+#ifndef mirai_cpp_events_MemberJoinEvent_hpp_H_
+#define mirai_cpp_events_MemberJoinEvent_hpp_H_
 
 #include "mirai/third-party/nlohmann/json.hpp"
 #include "mirai/defs/group_member.hpp"
@@ -39,4 +39,4 @@ namespace Cyan
 
 }
 
-#endif // !mirai_cpp_events_member_join_event_hpp_H_
+#endif
