+++ include/mirai/events/MemberLeaveEventKick.hpp
@@ -1,40 +1,35 @@
 #pragma once
-#ifndef mirai_cpp_events_group_mute_all_event_hpp_H_
-#define mirai_cpp_events_group_mute_all_event_hpp_H_
+#ifndef mirai_cpp_events_MemberLeaveEventKick_hpp_H_
+#define mirai_cpp_events_MemberLeaveEventKick_hpp_H_
 
 #include "mirai/third-party/nlohmann/json.hpp"
 #include "event_interface.hpp"
-#include "mirai/defs/group.hpp"
 #include "mirai/defs/group_member.hpp"
 
 namespace Cyan
 {
 	/**
-	 * \brief 群全体禁言事件
+	 * \brief 群成员被踢出群事件
 	 */
-	class GroupMuteAllEvent : public EventBase
+	class MemberLeaveEventKick : public EventBase
 	{
 	public:
-		bool Origin;
-		bool Current;
-		Group_t Group;
+		GroupMember Member;
 		GroupMember Operator;
 
 		static MiraiEvent GetMiraiEvent()
 		{
-			return MiraiEvent::GroupMuteAllEvent;
+			return MiraiEvent::MemberLeaveEventKick;
 		}
 
 		bool OperatorIsBot() const
 		{
 			return operator_is_null_;
 		}
-
+		
 		virtual bool Set(const json& j) override
 		{
-			this->Origin = j["origin"].get<bool>();
-			this->Current = j["current"].get<bool>();
-			this->Group.Set(j["group"]);
+			this->Member.Set(j["member"]);
 			if (!j["operator"].is_null())
 			{
 				this->Operator.Set(j["operator"]);
@@ -45,21 +40,18 @@ namespace Cyan
 		virtual json ToJson() const override
 		{
 			json j = json::object();
-			j["type"] = "GroupMuteAllEvent";
-			j["origin"] = this->Origin;
-			j["current"] = this->Current;
-			j["group"] = this->Group.ToJson();
+			j["type"] = "MemberLeaveEventKick";
+			j["member"] = this->Member.ToJson();
 			if (!operator_is_null_)
 				j["operator"] = this->Operator.ToJson();
 			else
 				j["operator"] = nullptr;
 			return j;
 		}
-
 	private:
 		bool operator_is_null_ = true;
 	};
 
 }
 
-#endif // !mirai_cpp_events_group_mute_all_event_hpp_H_
+#endif
