+++ include/mirai/events/MemberLeaveEventQuit.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#ifndef mirai_cpp_events_member_leave_quit_event_hpp_H_
-#define mirai_cpp_events_member_leave_quit_event_hpp_H_
+#ifndef mirai_cpp_events_MemberLeaveEventQuit_hpp_H_
+#define mirai_cpp_events_MemberLeaveEventQuit_hpp_H_
 
 #include "mirai/third-party/nlohmann/json.hpp"
 #include "mirai/defs/group_member.hpp"
@@ -37,4 +37,4 @@ namespace Cyan
 
 }
 
-#endif // !mirai_cpp_events_member_leave_quit_event_hpp_H_
+#endif
