+++ include/mirai/events/MemberMuteEvent.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#ifndef mirai_cpp_events_member_mute_event_hpp_H_
-#define mirai_cpp_events_member_mute_event_hpp_H_
+#ifndef mirai_cpp_events_MemberMuteEvent_hpp_H_
+#define mirai_cpp_events_MemberMuteEvent_hpp_H_
 
 #include "mirai/third-party/nlohmann/json.hpp"
 #include "event_interface.hpp"
@@ -58,4 +58,4 @@ namespace Cyan
 
 }
 
-#endif // !mirai_cpp_events_member_mute_event_hpp_H_
+#endif
