+++ include/mirai/events/NewFriendRequestEvent.hpp
@@ -1,6 +1,6 @@
 #pragma once
-#ifndef mirai_cpp_events_join_request_event_hpp_H_
-#define mirai_cpp_events_join_request_event_hpp_H_
+#ifndef mirai_cpp_events_NewFriendRequestEvent_hpp_H_
+#define mirai_cpp_events_NewFriendRequestEvent_hpp_H_
 
 #include "mirai/third-party/nlohmann/json.hpp"
 #include "mirai/defs/qq_types.hpp"
@@ -10,21 +10,20 @@
 namespace Cyan
 {
 	/**
-	 * \brief 新成员入群请求
+	 * \brief 新好友请求事件
 	 */
-	class EXPORTED MemberJoinRequestEvent : public EventBase
+	class EXPORTED NewFriendRequestEvent : public EventBase
 	{
 	public:
 		int64_t EventId;
 		QQ_t FromId;
 		GID_t GroupId;
-		string GroupName;
 		string Nick;
 		string Message;
 
 		static MiraiEvent GetMiraiEvent()
 		{
-			return MiraiEvent::MemberJoinRequestEvent;
+			return MiraiEvent::NewFriendRequestEvent;
 		}
 
 		bool Accept()
@@ -37,27 +36,16 @@ namespace Cyan
 			return Response(1, message);
 		}
 
-		bool Ignore(const string& message = "")
-		{
-			return Response(2, message);
-		}
-
 		bool RejectAndBlock(const string& message = "")
 		{
-			return Response(3, message);
-		}
-
-		bool IgnoreAndBlock(const string& message = "")
-		{
-			return Response(4, message);
+			return Response(2, message);
 		}
 
 		virtual bool Set(const json& j) override
 		{
 			this->EventId = j["eventId"].get<int64_t>();
-			this->FromId = QQ_t(j["fromId"].get<int64_t>());
-			this->GroupName = j["groupName"].get<string>();
-			this->GroupId = GID_t(j["groupId"].get<int64_t>());
+			this->FromId = (QQ_t)j["fromId"].get<int64_t>();
+			this->GroupId = (GID_t)j["groupId"].get<int64_t>();
 			this->Nick = j["nick"].get<string>();
 			this->Message = j["message"].get<string>();
 			return true;
@@ -65,11 +53,10 @@ namespace Cyan
 		virtual json ToJson() const override
 		{
 			json j = json::object();
-			j["type"] = "MemberJoinRequestEvent";
+			j["type"] = "NewFriendRequestEvent";
 			j["eventId"] = this->EventId;
 			j["fromId"] = (int64_t)this->FromId;
 			j["groupId"] = (int64_t)this->GroupId;
-			j["groupName"] = this->GroupName;
 			j["nick"] = this->Nick;
 			j["message"] = this->Message;
 			return j;
@@ -81,4 +68,4 @@ namespace Cyan
 
 }
 
-#endif // !mirai_cpp_events_join_request_event_hpp_H_
+#endif
